@@ -9,7 +9,7 @@ export '../../styles/icons.dart';
 /// 基于字体的图标集，可以通过 Icon 组件使用，也可以在其他组件中通过 `icon` 属性引用。
 class FlanIcon extends StatelessWidget {
   const FlanIcon({
-    Key key,
+    Key? key,
     this.iconName,
     this.iconUrl,
     this.dot = false,
@@ -18,62 +18,57 @@ class FlanIcon extends StatelessWidget {
     this.classPrefix = kFlanIconsFamily,
     this.badge,
     this.onClick,
-  })  : assert(dot != null),
-        super(key: key);
+  }) : super(key: key);
 
   const FlanIcon.name(
     this.iconName, {
-    Key key,
+    Key? key,
     this.dot = false,
     this.size,
     this.color,
     this.classPrefix = kFlanIconsFamily,
     this.badge,
     this.onClick,
-  })  : assert(iconName != null),
-        assert(dot != null),
-        this.iconUrl = null,
+  })  : this.iconUrl = null,
         super(key: key);
 
   const FlanIcon.url(
     this.iconUrl, {
-    Key key,
+    Key? key,
     this.dot = false,
     this.size,
     this.color,
     this.classPrefix = kFlanIconsFamily,
     this.badge,
     this.onClick,
-  })  : assert(iconUrl != null),
-        assert(dot != null),
-        this.iconName = null,
+  })  : this.iconName = null,
         super(key: key);
 
   // ****************** Props ******************
   /// 图标名称
-  final int iconName;
+  final int? iconName;
 
   /// 图片链接
-  final String iconUrl;
+  final String? iconUrl;
 
   /// 是否显示图标右上角小红点
   final bool dot;
 
   /// 图标右上角徽标的内容
-  final String badge;
+  final String? badge;
 
   /// 图标颜色
-  final Color color;
+  final Color? color;
 
   /// 图标大小
-  final double size;
+  final double? size;
 
   /// 类名前缀，用于使用自定义图标
   final String classPrefix;
 
   // ****************** Events ******************
   /// 点击图标时触发
-  final GestureTapCallback onClick;
+  final GestureTapCallback? onClick;
 
   // ****************** Slots ******************
 
@@ -102,17 +97,17 @@ class FlanIcon extends StatelessWidget {
 
     if (this.iconName != null) {
       return Icon(
-        IconData(this.iconName, fontFamily: this.classPrefix),
+        IconData(this.iconName!, fontFamily: this.classPrefix),
         color: this.color ?? iconTheme.color,
         size: iconSize,
       );
     }
 
-    final isNetWork = RegExp("^https?:\/\/").hasMatch(this.iconUrl);
+    final isNetWork = RegExp("^https?:\/\/").hasMatch(this.iconUrl!);
 
     if (isNetWork) {
       return Image.network(
-        this.iconUrl,
+        this.iconUrl!,
         color: this.color, //?? textStyle.color,
         width: iconSize,
         height: iconSize,
@@ -121,7 +116,7 @@ class FlanIcon extends StatelessWidget {
     }
 
     return Image.asset(
-      this.iconUrl,
+      this.iconUrl!,
       color: this.color, //?? textStyle.color,
       width: iconSize,
       height: iconSize,
